+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- `AppendStream` now caps each uncompressed NDJSON request at 8 MiB.
+
 ## 0.2.1 - 2026-08-20
 
 ### Changed